% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{AddonsConfig}
\alias{AddonsConfig}
\title{AddonsConfig Object}
\usage{
AddonsConfig(httpLoadBalancing = NULL, horizontalPodAutoscaling = NULL)
}
\arguments{
\item{httpLoadBalancing}{Configuration for the HTTP (L7) load balancing controller addon, which makes it easy to set up HTTP load balancers for services in a cluster}

\item{horizontalPodAutoscaling}{Configuration for the horizontal pod autoscaling feature, which increases or decreases the number of replica pods a replication controller has based on the resource usage of the existing pods}
}
\value{
AddonsConfig object
}
\description{
AddonsConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for the addons that can be automatically spun up in the cluster, enabling additional functionality.
}

