% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{AutoUpgradeOptions}
\alias{AutoUpgradeOptions}
\title{AutoUpgradeOptions Object}
\usage{
AutoUpgradeOptions(autoUpgradeStartTime = NULL, description = NULL)
}
\arguments{
\item{autoUpgradeStartTime}{[Output only] This field is set when upgrades are about to commence with the approximate start time for the upgrades, in [RFC3339](https://www}

\item{description}{[Output only] This field is set when upgrades are about to commence with the description of the upgrade}
}
\value{
AutoUpgradeOptions object
}
\description{
AutoUpgradeOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AutoUpgradeOptions defines the set of options for the user to control how the Auto Upgrades will proceed.
}
