% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{Cluster}
\alias{Cluster}
\title{Cluster Object}
\usage{
Cluster(name = NULL, description = NULL, initialNodeCount = NULL,
  nodeConfig = NULL, masterAuth = NULL, loggingService = NULL,
  monitoringService = NULL, network = NULL, clusterIpv4Cidr = NULL,
  addonsConfig = NULL, subnetwork = NULL, nodePools = NULL,
  locations = NULL, enableKubernetesAlpha = NULL, selfLink = NULL,
  zone = NULL, endpoint = NULL, initialClusterVersion = NULL,
  currentMasterVersion = NULL, currentNodeVersion = NULL,
  createTime = NULL, status = NULL, statusMessage = NULL,
  nodeIpv4CidrSize = NULL, servicesIpv4Cidr = NULL,
  instanceGroupUrls = NULL, currentNodeCount = NULL, expireTime = NULL)
}
\arguments{
\item{name}{The name of this cluster}

\item{description}{An optional description of this cluster}

\item{initialNodeCount}{The number of nodes to create in this cluster}

\item{nodeConfig}{Parameters used in creating the cluster's nodes}

\item{masterAuth}{The authentication information for accessing the master endpoint}

\item{loggingService}{The logging service the cluster should use to write logs}

\item{monitoringService}{The monitoring service the cluster should use to write metrics}

\item{network}{The name of the Google Compute Engine [network](/compute/docs/networks-and-firewalls#networks) to which the cluster is connected}

\item{clusterIpv4Cidr}{The IP address range of the container pods in this cluster, in [CIDR](http://en}

\item{addonsConfig}{Configurations for the various addons available to run in the cluster}

\item{subnetwork}{The name of the Google Compute Engine [subnetwork](/compute/docs/subnetworks) to which the cluster is connected}

\item{nodePools}{The node pools associated with this cluster}

\item{locations}{The list of Google Compute Engine [locations](/compute/docs/zones#available) in which the cluster's nodes should be located}

\item{enableKubernetesAlpha}{Kubernetes alpha features are enabled on this cluster}

\item{selfLink}{[Output only] Server-defined URL for the resource}

\item{zone}{[Output only] The name of the Google Compute Engine [zone](/compute/docs/zones#available) in which the cluster resides}

\item{endpoint}{[Output only] The IP address of this cluster's master endpoint}

\item{initialClusterVersion}{[Output only] The software version of the master endpoint and kubelets used in the cluster when it was first created}

\item{currentMasterVersion}{[Output only] The current software version of the master endpoint}

\item{currentNodeVersion}{[Output only] The current version of the node software components}

\item{createTime}{[Output only] The time the cluster was created, in [RFC3339](https://www}

\item{status}{[Output only] The current status of this cluster}

\item{statusMessage}{[Output only] Additional information about the current status of this cluster, if available}

\item{nodeIpv4CidrSize}{[Output only] The size of the address space on each node for hosting containers}

\item{servicesIpv4Cidr}{[Output only] The IP address range of the Kubernetes services in this cluster, in [CIDR](http://en}

\item{instanceGroupUrls}{[Output only] The resource URLs of [instance groups](/compute/docs/instance-groups/) associated with this cluster}

\item{currentNodeCount}{[Output only] The number of nodes currently in the cluster}

\item{expireTime}{[Output only] The time the cluster will be automatically deleted in [RFC3339](https://www}
}
\value{
Cluster object
}
\description{
Cluster Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google Container Engine cluster.
}
