% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_functions.R
\docType{package}
\name{container_googleAuthR}
\alias{container_googleAuthR}
\alias{container_googleAuthR-package}
\title{Google Container Engine API
Builds and manages clusters that run container-based applications, powered by open source Kubernetes technology.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:57:03
filename: /Users/mark/dev/R/autoGoogleAPI/googlecontainerv1.auto/R/container_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
