% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{CreateClusterRequest}
\alias{CreateClusterRequest}
\title{CreateClusterRequest Object}
\usage{
CreateClusterRequest(cluster = NULL)
}
\arguments{
\item{cluster}{A [cluster resource](/container-engine/reference/rest/v1/projects}
}
\value{
CreateClusterRequest object
}
\description{
CreateClusterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
CreateClusterRequest creates a cluster.
}
