% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{CreateNodePoolRequest}
\alias{CreateNodePoolRequest}
\title{CreateNodePoolRequest Object}
\usage{
CreateNodePoolRequest(nodePool = NULL)
}
\arguments{
\item{nodePool}{The node pool to create}
}
\value{
CreateNodePoolRequest object
}
\description{
CreateNodePoolRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
CreateNodePoolRequest creates a node pool for a cluster.
}

