% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{HorizontalPodAutoscaling}
\alias{HorizontalPodAutoscaling}
\title{HorizontalPodAutoscaling Object}
\usage{
HorizontalPodAutoscaling(disabled = NULL)
}
\arguments{
\item{disabled}{Whether the Horizontal Pod Autoscaling feature is enabled in the cluster}
}
\value{
HorizontalPodAutoscaling object
}
\description{
HorizontalPodAutoscaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration options for the horizontal pod autoscaling feature, which increases or decreases the number of replica pods a replication controller has based on the resource usage of the existing pods.
}

