% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{HttpLoadBalancing}
\alias{HttpLoadBalancing}
\title{HttpLoadBalancing Object}
\usage{
HttpLoadBalancing(disabled = NULL)
}
\arguments{
\item{disabled}{Whether the HTTP Load Balancing controller is enabled in the cluster}
}
\value{
HttpLoadBalancing object
}
\description{
HttpLoadBalancing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration options for the HTTP (L7) load balancing controller addon, which makes it easy to set up HTTP load balancers for services in a cluster.
}

