% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{ListClustersResponse}
\alias{ListClustersResponse}
\title{Google Container Engine API Objects 
Builds and manages clusters that run container-based applications, powered by open source Kubernetes technology.}
\usage{
ListClustersResponse(clusters = NULL, missingZones = NULL)
}
\arguments{
\item{clusters}{A list of clusters in the project in the specified zone, or across all ones}

\item{missingZones}{If any zones are listed here, the list of clusters returned may be missing those zones}
}
\value{
ListClustersResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:17:58
filename: /Users/mark/dev/R/autoGoogleAPI/googlecontainerv1.auto/R/container_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListClustersResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListClustersResponse is the result of ListClustersRequest.
}

