% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{ListNodePoolsResponse}
\alias{ListNodePoolsResponse}
\title{ListNodePoolsResponse Object}
\usage{
ListNodePoolsResponse(nodePools = NULL)
}
\arguments{
\item{nodePools}{A list of node pools for a cluster}
}
\value{
ListNodePoolsResponse object
}
\description{
ListNodePoolsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListNodePoolsResponse is the result of ListNodePoolsRequest.
}
