% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{ListOperationsResponse}
\alias{ListOperationsResponse}
\title{ListOperationsResponse Object}
\usage{
ListOperationsResponse(operations = NULL, missingZones = NULL)
}
\arguments{
\item{operations}{A list of operations in the project in the specified zone}

\item{missingZones}{If any zones are listed here, the list of operations returned may be missing the operations from those zones}
}
\value{
ListOperationsResponse object
}
\description{
ListOperationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListOperationsResponse is the result of ListOperationsRequest.
}

