% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{MasterAuth}
\alias{MasterAuth}
\title{MasterAuth Object}
\usage{
MasterAuth(username = NULL, password = NULL, clusterCaCertificate = NULL,
  clientCertificate = NULL, clientKey = NULL)
}
\arguments{
\item{username}{The username to use for HTTP basic authentication to the master endpoint}

\item{password}{The password to use for HTTP basic authentication to the master endpoint}

\item{clusterCaCertificate}{[Output only] Base64-encoded public certificate that is the root of trust for the cluster}

\item{clientCertificate}{[Output only] Base64-encoded public certificate used by clients to authenticate to the cluster endpoint}

\item{clientKey}{[Output only] Base64-encoded private key used by clients to authenticate to the cluster endpoint}
}
\value{
MasterAuth object
}
\description{
MasterAuth Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The authentication information for accessing the master endpoint. Authentication can be done using HTTP basic auth or using client certificates.
}
