% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{NodeConfig.labels}
\alias{NodeConfig.labels}
\title{NodeConfig.labels Object}
\usage{
NodeConfig.labels()
}
\value{
NodeConfig.labels object
}
\description{
NodeConfig.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The map of Kubernetes labels (key/value pairs) to be applied to each node. These will added in addition to any default label(s) that Kubernetes may apply to the node. In case of conflict in label keys, the applied set may differ depending on the Kubernetes version -- it's best to assume the behavior is undefined and conflicts should be avoided. For more information, including usage and the valid values, see: http://kubernetes.io/v1.1/docs/user-guide/labels.html
}
\seealso{
Other NodeConfig functions: \code{\link{NodeConfig.metadata}},
  \code{\link{NodeConfig}}
}
