% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{NodeConfig.metadata}
\alias{NodeConfig.metadata}
\title{NodeConfig.metadata Object}
\usage{
NodeConfig.metadata()
}
\value{
NodeConfig.metadata object
}
\description{
NodeConfig.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata key/value pairs assigned to instances in the cluster. Keys must conform to the regexp [a-zA-Z0-9-_]+ and be less than 128 bytes in length. These are reflected as part of a URL in the metadata server. Additionally, to avoid ambiguity, keys must not conflict with any other metadata keys for the project or be one of the four reserved keys: 'instance-template', 'kube-env', 'startup-script', and 'user-data' Values are free-form strings, and only have meaning as interpreted by the image running in the instance. The only restriction placed on them is that each value's size must be less than or equal to 32 KB. The total size of all keys and values must be less than 512 KB.
}
\seealso{
Other NodeConfig functions: \code{\link{NodeConfig}}
}

