% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{NodeConfig}
\alias{NodeConfig}
\title{NodeConfig Object}
\usage{
NodeConfig(NodeConfig.metadata = NULL, NodeConfig.labels = NULL,
  machineType = NULL, diskSizeGb = NULL, oauthScopes = NULL,
  serviceAccount = NULL, metadata = NULL, imageType = NULL,
  labels = NULL, localSsdCount = NULL, tags = NULL, preemptible = NULL)
}
\arguments{
\item{NodeConfig.metadata}{The \link{NodeConfig.metadata} object or list of objects}

\item{NodeConfig.labels}{The \link{NodeConfig.labels} object or list of objects}

\item{machineType}{The name of a Google Compute Engine [machine type](/compute/docs/machine-types) (e}

\item{diskSizeGb}{Size of the disk attached to each node, specified in GB}

\item{oauthScopes}{The set of Google API scopes to be made available on all of the node VMs under the 'default' service account}

\item{serviceAccount}{The Google Cloud Platform Service Account to be used by the node VMs}

\item{metadata}{The metadata key/value pairs assigned to instances in the cluster}

\item{imageType}{The image type to use for this node}

\item{labels}{The map of Kubernetes labels (key/value pairs) to be applied to each node}

\item{localSsdCount}{The number of local SSD disks to be attached to the node}

\item{tags}{The list of instance tags applied to all nodes}

\item{preemptible}{Whether the nodes are created as preemptible VM instances}
}
\value{
NodeConfig object
}
\description{
NodeConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters that describe the nodes in a cluster.
}
\seealso{
Other NodeConfig functions: \code{\link{NodeConfig.labels}},
  \code{\link{NodeConfig.metadata}}
}
