% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{NodeManagement}
\alias{NodeManagement}
\title{NodeManagement Object}
\usage{
NodeManagement(autoUpgrade = NULL, upgradeOptions = NULL)
}
\arguments{
\item{autoUpgrade}{Whether the nodes will be automatically upgraded}

\item{upgradeOptions}{Specifies the Auto Upgrade knobs for the node pool}
}
\value{
NodeManagement object
}
\description{
NodeManagement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
NodeManagement defines the set of node management services turned on for the node pool.
}
