% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{NodePool}
\alias{NodePool}
\title{NodePool Object}
\usage{
NodePool(name = NULL, config = NULL, initialNodeCount = NULL,
  selfLink = NULL, version = NULL, instanceGroupUrls = NULL,
  status = NULL, statusMessage = NULL, autoscaling = NULL,
  management = NULL)
}
\arguments{
\item{name}{The name of the node pool}

\item{config}{The node configuration of the pool}

\item{initialNodeCount}{The initial node count for the pool}

\item{selfLink}{[Output only] Server-defined URL for the resource}

\item{version}{[Output only] The version of the Kubernetes of this node}

\item{instanceGroupUrls}{[Output only] The resource URLs of [instance groups](/compute/docs/instance-groups/) associated with this node pool}

\item{status}{[Output only] The status of the nodes in this pool instance}

\item{statusMessage}{[Output only] Additional information about the current status of this node pool instance, if available}

\item{autoscaling}{Autoscaler configuration for this NodePool}

\item{management}{NodeManagement configuration for this NodePool}
}
\value{
NodePool object
}
\description{
NodePool Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
NodePool contains the name and configuration for a cluster's node pool. Node pools are a set of nodes (i.e. VM's), with a common configuration and specification, under the control of the cluster master. They may have a set of Kubernetes labels applied to them, which may be used to reference them during pod scheduling. They may also be resized up or down, to accommodate the workload.
}
