% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{NodePoolAutoscaling}
\alias{NodePoolAutoscaling}
\title{NodePoolAutoscaling Object}
\usage{
NodePoolAutoscaling(enabled = NULL, minNodeCount = NULL,
  maxNodeCount = NULL)
}
\arguments{
\item{enabled}{Is autoscaling enabled for this node pool}

\item{minNodeCount}{Minimum number of nodes in the NodePool}

\item{maxNodeCount}{Maximum number of nodes in the NodePool}
}
\value{
NodePoolAutoscaling object
}
\description{
NodePoolAutoscaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
NodePoolAutoscaling contains information required by cluster autoscaler to adjust the size of the node pool to the current cluster usage.
}
