% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(name = NULL, zone = NULL, operationType = NULL, status = NULL,
  detail = NULL, statusMessage = NULL, selfLink = NULL,
  targetLink = NULL)
}
\arguments{
\item{name}{The server-assigned ID for the operation}

\item{zone}{The name of the Google Compute Engine [zone](/compute/docs/zones#available) in which the operation is taking place}

\item{operationType}{The operation type}

\item{status}{The current status of the operation}

\item{detail}{Detailed operation progress, if available}

\item{statusMessage}{If an error has occurred, a textual description of the error}

\item{selfLink}{Server-defined URL for the resource}

\item{targetLink}{Server-defined URL for the target of the operation}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This operation resource represents operations that may have happened or are happening on the cluster. All fields are output only.
}

