% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_functions.R
\name{projects.zones.getServerconfig}
\alias{projects.zones.getServerconfig}
\title{Returns configuration info about the Container Engine service.}
\usage{
projects.zones.getServerconfig(projectId, zone)
}
\arguments{
\item{projectId}{The Google Developers Console [project ID or project number](https://support}

\item{zone}{The name of the Google Compute Engine [zone](/compute/docs/zones#available) to return operations for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/container-engine/}{Google Documentation}
}
