% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{RollbackNodePoolUpgradeRequest}
\alias{RollbackNodePoolUpgradeRequest}
\title{RollbackNodePoolUpgradeRequest Object}
\usage{
RollbackNodePoolUpgradeRequest()
}
\value{
RollbackNodePoolUpgradeRequest object
}
\description{
RollbackNodePoolUpgradeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
RollbackNodePoolUpgradeRequest rollbacks the previously Aborted or Failed NodePool upgrade. This will be an no-op if the last upgrade successfully completed.
}
