% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{SetNodePoolManagementRequest}
\alias{SetNodePoolManagementRequest}
\title{SetNodePoolManagementRequest Object}
\usage{
SetNodePoolManagementRequest(management = NULL)
}
\arguments{
\item{management}{NodeManagement configuration for the node pool}
}
\value{
SetNodePoolManagementRequest object
}
\description{
SetNodePoolManagementRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SetNodePoolManagementRequest sets the node management properties of a node pool.
}
