% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{UpdateClusterRequest}
\alias{UpdateClusterRequest}
\title{UpdateClusterRequest Object}
\usage{
UpdateClusterRequest(update = NULL)
}
\arguments{
\item{update}{A description of the update}
}
\value{
UpdateClusterRequest object
}
\description{
UpdateClusterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
UpdateClusterRequest updates the settings of a cluster.
}
