% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{ClusterUpdate}
\alias{ClusterUpdate}
\title{ClusterUpdate Object}
\usage{
ClusterUpdate(desiredNodeVersion = NULL, desiredMonitoringService = NULL,
  desiredAddonsConfig = NULL, desiredNodePoolId = NULL,
  desiredMasterVersion = NULL)
}
\arguments{
\item{desiredNodeVersion}{The Kubernetes version to change the nodes to (typically an upgrade)}

\item{desiredMonitoringService}{The monitoring service the cluster should use to write metrics}

\item{desiredAddonsConfig}{Configurations for the various addons available to run in the cluster}

\item{desiredNodePoolId}{The node pool to be upgraded}

\item{desiredMasterVersion}{The Kubernetes version to change the master to}
}
\value{
ClusterUpdate object
}
\description{
ClusterUpdate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ClusterUpdate describes an update to the cluster. Exactly one update can be applied to a cluster with each request, so at most one field can be provided.
}

