% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_objects.R
\name{ServerConfig}
\alias{ServerConfig}
\title{ServerConfig Object}
\usage{
ServerConfig(defaultClusterVersion = NULL, validNodeVersions = NULL,
  defaultImageFamily = NULL, validImageFamilies = NULL)
}
\arguments{
\item{defaultClusterVersion}{Version of Kubernetes the service deploys by default}

\item{validNodeVersions}{List of valid node upgrade target versions}

\item{defaultImageFamily}{Default image family}

\item{validImageFamilies}{List of valid image families}
}
\value{
ServerConfig object
}
\description{
ServerConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Container Engine service configuration.
}

